requirements.txt
pandas
sqlalchemy
azure-storage-blob
pyodbc
python-dotenv